+++ api/tests/test_api_video.py
@@ -1,12 +1,14 @@
+from django.test import client
 from rest_framework import status
 from rest_framework.test import APIClient
 from django.contrib.gis.geos import Point
 from django.core.files.uploadedfile import SimpleUploadedFile
 from .test_setup_db import InitTestCase
 # Create your tests here.
-from api.models import Site, Recipe, CrossSection
+from api.models import Site, Recipe, CrossSection, Video
 from users.models import User, Institute
-from datetime import datetime
+from datetime import datetime, timedelta
+import copy
 import json
 import os
 import requests
@@ -16,9 +18,13 @@ from .test_api_recipe import recipe
 from .test_api_profile import profile
 
 
+
+
 video_sample_url = "https://raw.githubusercontent.com/localdevices/pyorc/main/examples/ngwerere/ngwerere_20191103.mp4"
 camconfig_url = "https://raw.githubusercontent.com/localdevices/pyorc/main/examples/ngwerere/ngwerere.json"
 image_sample_url = "https://raw.githubusercontent.com/localdevices/pyorc/main/docs/ngwerere.jpg"
+result_2d_sample_url = "https://raw.githubusercontent.com/localdevices/pyorc/main/examples/ngwerere_piv.nc"
+result_2d_mask_url = "https://raw.githubusercontent.com/localdevices/pyorc/main/examples/ngwerere_masked.nc"
 
 
 def prep_video_sample(video_sample_url):
@@ -62,6 +68,23 @@ def prep_image_sample(image_sample_url):
     }
     return msg
 
+def prep_results_sample():
+    src_urls = [result_2d_sample_url, result_2d_mask_url]
+    dst_fns = ["2d.nc", "2d_mask.nc"]
+    # prepare basic fields of msg content
+    msg = {
+        "timestamp": datetime.now().strftime("%Y-%m-%dT%H:%M:%SZ"),
+        "video_config": 1,
+    }
+
+    for url, fn in zip(src_urls, dst_fns):
+        print(f"Downloading {url}")
+        r = requests.get(url)
+        obj = r.content
+        result_file = SimpleUploadedFile(fn, obj, content_type="application/x-netcdf")
+        msg[f"result_{os.path.split(fn)[-1].replace('.nc', '')}"] = result_file
+    return msg
+
 
 
 def camconfig(camconfig_url):
@@ -72,6 +95,16 @@ def camconfig(camconfig_url):
 camera_config = camconfig(camconfig_url)
 video_sample = prep_video_sample(video_sample_url)
 image_sample = prep_image_sample(image_sample_url)
+results_sample = prep_results_sample()
+
+# test also patching an already existing video with results
+results_sample_patch = {
+    "result_2d": copy.deepcopy(results_sample["result_2d"]),
+    "result_2d_mask": copy.deepcopy(results_sample["result_2d_mask"]),
+    "log_file": SimpleUploadedFile("orc.log", b"Some log content2", content_type="text/plain")
+}
+
+
 no_files_sample = prep_no_files_sample()
 
 camera_config_form = {
@@ -96,11 +129,12 @@ class VideoViewTests(InitTestCase):
         site = Site.objects.create(name="ngwerere", geom=Point(28.329686, -15.334151), creator=user, institute=institute)
         Recipe.objects.create(name="ngwerere_recipe", data=recipe, creator=user, institute=institute)
         CrossSection.objects.create(name="some_cross_section", features=profile, site=site, creator=user)
+        self.test_timestamp = datetime.now()
 
     def tearDown(self):
-        pass
+        Video.objects.all().delete()
 
-    def test_add_video(self):
+    def test_add_video_patch_results_delete(self):
         client = APIClient()
         client.login(username='user@institute1.com', password='test1234')
         # create a camera config on site
@@ -145,11 +179,33 @@ class VideoViewTests(InitTestCase):
             # follow=True
         )
         self.assertEqual(r.status_code, status.HTTP_403_FORBIDDEN)
-
+        client.logout()
+        # log back in as original posting user and patch the video with results
+        client.login(username='user@institute1.com', password='test1234')
+        # now patch the video with new information
+        r = client.patch(
+            '/api/site/1/video/1/',
+            data=results_sample_patch,
+            # follow=True
+        )
+        self.assertEqual(r.status_code, status.HTTP_200_OK)
+        # also check presence of result files
+        r = client.get("/api/site/1/video/1/result_2d/")
+        self.assertEqual(r.status_code, status.HTTP_200_OK)
+        # check the content of the log file
+        r = client.get("/api/site/1/video/1/log_file/")
+        self.assertEqual(r.status_code, status.HTTP_200_OK)
+        self.assertIn("Some log content", r.content.decode())
+        # delete
+        r = client.delete("/api/site/1/video/1/")
+        self.assertEqual(r.status_code, status.HTTP_204_NO_CONTENT)
 
     def test_add_image(self):
+        self.test_timestamp = datetime.now() + timedelta(seconds=1)
         client = APIClient()
         client.login(username='user@institute1.com', password='test1234')
+        image_sample = prep_image_sample(image_sample_url)
+
         # create a camera config on site
         r = client.post(
             '/api/site/1/cameraconfig/',
@@ -170,7 +226,10 @@ class VideoViewTests(InitTestCase):
         self.assertIsNotNone(r.data["image"])
         self.assertIsNone(r.data["file"])
         self.assertIsNotNone(r.data["thumbnail"])
-
+        # there should not be any result files yet, check this
+        r = client.get("/api/site/1/video/1/result_2d/")
+        self.assertEqual(r.status_code, status.HTTP_404_NOT_FOUND)
+        assert "No result file" in r.text
 
     def test_add_no_files(self):
         client = APIClient()
@@ -196,7 +255,34 @@ class VideoViewTests(InitTestCase):
         self.assertIsNone(r.data["image"])
         self.assertIsNone(r.data["thumbnail"])
 
-
+    def test_add_results(self):
+        client = APIClient()
+        client.login(username='user@institute1.com', password='test1234')
+        # create a camera config on site
+        r = client.post(
+            '/api/site/1/cameraconfig/',
+            camera_config_form
+        )
+        self.assertEqual(r.status_code, status.HTTP_201_CREATED)
+        r = client.post('/api/site/1/videoconfig/', video_config_form)
+        self.assertEqual(r.status_code, status.HTTP_201_CREATED)
+        # post a video with only the result image instead of full video
+        r = client.post(
+            "/api/video/",
+            data=results_sample
+        )
+        self.assertEqual(r.status_code, status.HTTP_201_CREATED)
+        r = client.get("/api/site/1/video/1/")
+        self.assertEqual(r.status_code, status.HTTP_200_OK)
+        # check if image is there, but file not
+        self.assertIsNone(r.data["image"])
+        self.assertIsNone(r.data["file"])
+        self.assertIsNone(r.data["thumbnail"])
+        # there should be a result in result_2d but not in result_1d as that was not included
+        r = client.get("/api/site/1/video/1/result_2d/")
+        self.assertEqual(r.status_code, status.HTTP_200_OK)
+        r = client.get("/api/site/1/video/1/result_1d/")
+        self.assertEqual(r.status_code, status.HTTP_404_NOT_FOUND)